+++ src/main/java/ru/dpopkov/tasktracker/Main.java
@@ -13,17 +13,23 @@ import ru.dpopkov.tasktracker.ui.actions.ActionPack;
 public class Main {
 
     public static void main(String[] args) {
+        TaskTracker taskTracker = initTaskTracker();
+        UiCycle uiCycle = initUiCycle(taskTracker);
+        uiCycle.run();
+    }
+
+    private static TaskTracker initTaskTracker() {
         ProjectService projectService = new ProjectServiceImpl(new SimpleProjectRepository());
         UserService userService = new UserServiceImpl(new SimpleUserRepository());
-        TaskTracker taskTracker = new TaskTracker(userService, projectService);
+        return new TaskTracker(userService, projectService);
+    }
 
+    private static UiCycle initUiCycle(TaskTracker taskTracker) {
         UiInput input = new ConsoleUiInput();
         UiOutput output = new ConsoleUiOutput();
         ActionPack pack = new ActionPack(input, output, taskTracker);
         Actions actions = new Actions(pack);
-
         Menu menu = new TextualMenu(actions, output, input);
-        UiCycle uiCycle = new UiCycle(actions, menu);
-        uiCycle.run();
+        return new UiCycle(actions, menu);
     }
 }